Deprecate compression option
